Modified kernel exposure
